requirements.txt
torch~=1.13.1
torchvision~=0.14.1
matplotlib~=3.7.0
numpy~=1.24.2
Pillow~=9.4.0
opencv-python~=4.5.5.62
tqdm~=4.64.1
wandb~=0.13.10
torchinfo~=1.7.2
imgaug~=0.4.0
pandas~=1.5.3